FIFO (named pipe) demo.
This is an experiment to send data over a named pipe, and allow for
killing either end. In Android, services may die if nothing's using
them; it's important that the ANT <-> HHAL and ANT <-> SLINK HALs
don't kill each other if one goes down.

In this demo, run `make`, then run `fifo_server` and `fifo_client` in
different terminals.

Observations:
 - The client and server wait for each other. Neither proceeds past the
   open() call until both are ready.
 - If the server dies, the client sits ready. When restarted, the server
   detects that client is present and immediately resumes sending data.
 - If the client dies, the server is killed with SIGPIPE (141). It'd be
   possible to handle that, but a little complicated.
 - Pipes are always 1-1. We'd like a simple way for the ANT HAL to
   broadcast to all interested HALs. Going to look into broadcast
   sockets.